pyproject.toml
[project]
name = "pdf-tei-editor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi>=0.116.2",
    "flask>=3.1.0",
    "llamore>=0.2.0",
    "lxml>=5.3.1",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "routes>=2.5.1",
    "uvicorn[standard]>=0.36.0",
    "waitress>=3.0.2",
    "webdav4[fsspec]>=0.10.0",
    "xmlschema>=3.4.5",
]

[tool.uv.sources]
llamore = { git = "https://github.com/mpilhlt/llamore.git" }

[dependency-groups]
dev = [
    "cheroot>=10.0.0",
    "lxml-stubs>=0.5.1",
    "mypy>=1.18.2",
    "pydeps>=3.0.1",
    "pytest>=8.4.0",
    "pytest-tap>=3.3",
    "wsgidav>=4.3.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
markers = [
  "integration: marks tests as integration tests that check actual project structure (deselect with '-m \"not integration\"')"
]
